ReservedFunctionNames: bugfix for PHPCS 2.3.x
Loosely related to 751.

When this sniff would run on PHPCS 2.3.x and the function being examined would be in an anonymous class, the class name used in the error message would be incorrect as the `getDeclarationName()` method would use the first T_STRING found, which in the case of an anonymous class would be the name of the first constant or function within the anonymous class.

Additionally, the upstream `processTokenWithinScope()` method is called for each valid scope a function is in, so for a function in a nested anonymous class, the method would be called twice and the error would also be thrown twice, though with two different class names - both wrong.

This PR fixes both these issues.

Includes additional unit test for this issue, though as the class/method names being passed to the error message are not covered by the unit tests and the unit tests only test that the error is found on a line, not how _often_ the same error is found on a line, the issue can only be seen when testing this manually (for now).